SWDEV-492563 - Fix Ocl issues
1. Fix LDSSize type to be uint32_t.
2. Prevent clWaitForEvents running on complete events whose
   HostQueue have been destructed.



[ROCm/clr commit: 82dff9a67d74345e35164c68ca1dfe9fae963eac]

+++ rocclr/device/pal/palkernel.cpp
@@ -441,7 +441,8 @@ hsa_kernel_dispatch_packet_t* HSAILKernel::loadArguments(
         WriteAqlArgAt(hidden_arguments, gpu.hsaQueueMem()->vmAddress(), it.size_, it.offset_);
         break;
       case amd::KernelParameterDescriptor::HiddenDynamicLdsSize:
-        WriteAqlArgAt(hidden_arguments, ldsAddress - ldsSize(), it.size_, it.offset_);
+        WriteAqlArgAt(hidden_arguments, static_cast<uint32_t>(ldsAddress - ldsSize()),
+               it.size_, it.offset_);
         break;
     }
   }
